requirements.txt
pandas~=1.4.4
numpy
python-binance